test: move the onnxruntime import pin to tests/test_import.py
The pin does not need onnx, but tests/feature/test_lightglue_onnx.py is
gated by pytest.importorskip("onnx"), so it ran only where onnx happened
to be installed. tests/test_import.py already hosts the subprocess import
guards, runs in every default CI leg, and gives the child interpreter the
same timeout=300 as its neighbours.
