+++ tests/test_import.py
@@ -18,6 +18,9 @@
 import subprocess
 import sys
 import textwrap
+from pathlib import Path
+
+import pytest
 
 
 def test_import_without_jit_script_deprecation():
@@ -98,3 +101,33 @@ def test_import_emits_no_warnings():
         timeout=300,
     )
     assert result.returncode == 0, f"importing kornia emitted a warning:\n{result.stderr}"
+
+
+def test_import_kornia_does_not_import_onnxruntime():
+    """``import kornia`` must not eagerly load onnxruntime, even when it is installed.
+
+    ``kornia.feature.lightglue_onnx`` used to import onnxruntime at module level, so a plain
+    ``import kornia`` loaded the ORT native extension for a class nobody had instantiated.
+    Every ONNX consumer now goes through the ``LazyLoader`` handles in
+    ``kornia.core.external``; this pin keeps it that way. See
+    https://github.com/kornia/kornia/issues/4260.
+
+    Skipped when onnxruntime is not installed, otherwise the assertion is vacuous. The child
+    interpreter runs with the repository root as its working directory so it imports the same
+    ``kornia`` as this session, even when kornia is installed editable from another checkout.
+    """
+    pytest.importorskip("onnxruntime")
+
+    import kornia
+
+    repo_root = Path(kornia.__file__).resolve().parents[1]
+    result = subprocess.run(
+        [sys.executable, "-c", "import kornia, sys; print('onnxruntime' in sys.modules)"],
+        cwd=repo_root,
+        capture_output=True,
+        text=True,
+        check=False,
+        timeout=300,
+    )
+    assert result.returncode == 0, f"importing kornia failed:\n{result.stderr}"
+    assert result.stdout.strip() == "False", f"import kornia loaded onnxruntime:\n{result.stdout}{result.stderr}"
